+++ tell_me_your_secrets/utils.py
@@ -1,5 +1,6 @@
 import math
 import pathlib
+from typing import Optional
 
 import yaml